requirements.txt
django-3.1
djangorestframework-3.11.1
markdown
django-filter
mysqlclient